requirements.txt
configparser
alpaca_trade_api 
pandas
numpy
yfinance==0.2.29
tensorflow==2.11.0
keras 
scikit-learn
mplfinance 
matplotlib 
ccxt
hmmlearn
requests
beautifulsoup4
statsmodels
#sqlite3
arch
ta
tqdm
kucoin-python
kucoin-futures-python
mysql
#abc
#talib 